// PetHealth_FE/src/models/features.ts
export interface PetHistoryDto {
  maLichHen: number;
  maThuCung: number;
  tenThuCung: string;
  tenDichVu: string;
  ngayHen: string;
  gioHen: string;
  trangThai: string;
  ghiChu?: string | null;
}

export interface PetVisitImageDto {
  maAnh: number;
  maLichHen: number;
  maThuCung: number;
  loaiAnh: string;
  anhUrl: string;
  ghiChu?: string | null;
  ngayTaiLen: string;
}